+++ android/src/oldarch/SmileIdSpec.kt
@@ -8,22 +8,13 @@ import com.facebook.react.bridge.ReadableMap
 abstract class SmileIdSpec internal constructor(context: ReactApplicationContext) :
   ReactContextBaseJavaModule(context) {
 
-  abstract fun initializeWithApiKey(
-    apiKey: String,
-    config: ReadableMap,
+  abstract fun initialize(
     useSandBox: Boolean,
     enableCrashReporting: Boolean,
+    config: ReadableMap?,
+    apiKey: String?,
     promise: Promise
   )
-
-  abstract fun initializeWithConfig(
-    config: ReadableMap,
-    useSandBox: Boolean,
-    enableCrashReporting: Boolean,
-    promise: Promise
-  )
-
-  abstract fun initialize(useSandBox: Boolean, promise: Promise)
 
   abstract fun setCallbackUrl(callbackUrl: String, promise: Promise)