+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 10.2.3
 * Fixed missing `didSubmitBiometricKycJob` flag in BiometricKyc response on android
+* Added `apiKey` and `config` missing to allow multiple initialization options see (https://docs.usesmileid.com/integration-options/mobile/getting-started)
+* Fix document capture restore optional parameters and showing or hiding instruction and confirmation screen
+* Bump android to 10.4.2 (https://github.com/smileidentity/android/releases/tag/v10.4.2)
+* Bump iOS to 10.3.4 (https://github.com/smileidentity/ios/releases/tag/v10.3.4)
 
 ## 10.2.2
 * Fixed setting `kotlinVersion` and `kotlinCompilerExtensionVersion`